refactor: extract useStreamdownConfig hook and formatRelativeTime util
Eliminate duplicated Streamdown code-block config (~20 lines) shared
between MessageBubble and StreamingMessage into a single reusable hook.
Extract the duplicated relative-time formatting logic from MessageBubble
and ContextSelector into a shared formatRelativeTime util.
